requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
supabase==2.17.0

# AI
google-generativeai==0.3.2

# Data validation
pydantic==2.8.0

# HTTP and utilities
httpx==0.25.0
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0

# Templates for UI
jinja2==3.1.2
markupsafe==2.1.3

# Date handling
python-dateutil==2.8.2

# Essential utilities
typing-extensions==4.12.0
